pyproject.toml
[tool.poetry]
name = "backend"
version = "1.0.0"
description = "ProDCube Backend"
authors = ["ProDCube <prodcube@prodcube.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = { extras = ["requests"], version = "^0.73.0" }
uvicorn = "^0.17.4"
pydantic = { extras = ["dotenv"], version = "^1.9.0" }
python-dotenv = "^0.19.2"
Jinja2 = "^3.0.3"
python-multipart = "^0.0.5"
passlib = "^1.7.4"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
bcrypt = "^3.2.0"
itsdangerous = "^2.0.1"
httpx = "^0.22.0"
pytest-cov = "^3.0.0"
requests = "^2.27.1"
trio = "^0.19.0"
mongoengine = "^0.23.1"
dnspython = "^2.2.0"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"
pylint-quotes = "^0.2.3"
isort = "^5.10.1"

[tool.isort]
indent = 4
src_paths = ["./"]
skip_gitignore = true
combine_as_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "Tests",
]
asyncio_mode = "strict"
python_files = "Test*.py"